break out Evacuate into new file
